+++ src/ast/nodes/shared/MethodTypes.ts
@@ -1,9 +1,5 @@
 import type { HasEffectsContext } from '../../ExecutionContext';
-import type {
-	NodeInteraction,
-	NodeInteractionCalled,
-	NodeInteractionWithThisArgument
-} from '../../NodeInteractions';
+import type { NodeInteraction, NodeInteractionCalled } from '../../NodeInteractions';
 import {
 	INTERACTION_ACCESSED,
 	INTERACTION_CALLED,
@@ -37,12 +33,12 @@ export class Method extends ExpressionEntity {
 		super();
 	}
 
-	deoptimizeThisOnInteractionAtPath(
-		{ type, thisArg }: NodeInteractionWithThisArgument,
+	deoptimizeArgumentsOnInteractionAtPath(
+		{ type, thisArg }: NodeInteraction,
 		path: ObjectPath
 	): void {
 		if (type === INTERACTION_CALLED && path.length === 0 && this.description.mutatesSelfAsArray) {
-			thisArg.deoptimizePath(UNKNOWN_INTEGER_PATH);
+			thisArg?.deoptimizePath(UNKNOWN_INTEGER_PATH);
 		}
 	}