+++ src/ast/nodes/shared/ObjectPrototype.ts
@@ -1,9 +1,8 @@
-import type { NodeInteraction, NodeInteractionWithThisArgument } from '../../NodeInteractions';
+import type { NodeInteraction } from '../../NodeInteractions';
 import { INTERACTION_CALLED } from '../../NodeInteractions';
 import type { ObjectPath, ObjectPathKey } from '../../utils/PathTracker';
-import { UNKNOWN_PATH } from '../../utils/PathTracker';
 import type { LiteralValueOrUnknown } from './Expression';
-import { ExpressionEntity, UnknownValue } from './Expression';
+import { deoptimizeInteraction, ExpressionEntity, UnknownValue } from './Expression';
 import {
 	METHOD_RETURNS_BOOLEAN,
 	METHOD_RETURNS_STRING,
@@ -20,12 +19,9 @@ const isInteger = (property: ObjectPathKey): boolean =>
 // will improve tree-shaking for out-of-bounds array properties
 const OBJECT_PROTOTYPE_FALLBACK: ExpressionEntity =
 	new (class ObjectPrototypeFallbackExpression extends ExpressionEntity {
-		deoptimizeThisOnInteractionAtPath(
-			{ type, thisArg }: NodeInteractionWithThisArgument,
-			path: ObjectPath
-		): void {
-			if (type === INTERACTION_CALLED && path.length === 1 && !isInteger(path[0])) {
-				thisArg.deoptimizePath(UNKNOWN_PATH);
+		deoptimizeArgumentsOnInteractionAtPath(interaction: NodeInteraction, path: ObjectPath): void {
+			if (interaction.type === INTERACTION_CALLED && path.length === 1 && !isInteger(path[0])) {
+				deoptimizeInteraction(interaction);
 			}
 		}