+++ test/form/samples/argument-deoptimization/no-default-deoptimization/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'does not deoptimize call arguments by default'
+};