+++ test/form/samples/argument-deoptimization/no-global-call-deoptimization/main.js
@@ -1,0 +1,8 @@
+const obj = { foo: true, noEffect() {} };
+
+console.log(obj);
+
+// removed
+obj.noEffect();
+
+console.log(obj.foo ? 'OK' : 'FAIL');