+++ test/function/samples/argument-deoptimization/global-calls/_config.js
@@ -1,0 +1,8 @@
+module.exports = {
+	description: 'tracks argument mutations of calls to globals',
+	options: {
+		treeshake: {
+			tryCatchDeoptimization: false
+		}
+	}
+};