+++ test/function/samples/argument-deoptimization/parameter-access-mutation/main.js
@@ -1,0 +1,14 @@
+const obj = {
+	get mutator() {
+		this.mutated = true;
+	},
+	mutated: false
+};
+
+function updateObj(target) {
+	target.mutator;
+}
+
+updateObj(obj);
+
+assert.ok(obj.mutated ? true : false);