+++ test/function/samples/argument-deoptimization/return-function-known/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'tracks mutations of function parameters by known returned functions'
+};