+++ test/function/samples/argument-deoptimization/unknown-global-assignment/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'tracks mutations of assignments to unknown globals'
+};